flask_app/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
PyJWT==2.8.0
Werkzeug==3.0.1
